requirements.txt
# --- Основные компоненты Airflow ---
apache-airflow==2.10.5
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-amazon==8.6.0

# --- Аналитика и хранилища ---
clickhouse-connect==0.7.17
sqlalchemy==1.4.36
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
duckdb==1.2.2   # оставим как дополнительное in-memory хранилище (опционально)

# --- Машинное обучение и прогноз ---
scikit-learn==1.4.2
prophet==1.1.5
xgboost==2.0.3

# --- Визуализация и интерфейсы ---
streamlit==1.33.0
plotly==5.19.0

# --- Хранилище файлов и модели ---
boto3==1.34.28
minio==7.2.7
mlflow==2.8.1

# --- Утилиты и API ---
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
